postgres_to_es/requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
flake8==6.1.0
idna==3.6
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.1.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.1
